requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.41
psycopg2-binary==2.9.9
Flask-Migrate==4.1.0
python-dotenv==1.1.0
Werkzeug==3.1.3
gunicorn==23.0.0